requirements.txt
accelerate==1.1.1
aioboto3==13.2.0
aiobotocore==2.15.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.3
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
anytree==2.12.1
appdirs==1.4.4
argcomplete==1.10.3
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
autograd==1.7.0
backoff==2.2.1
beartype==0.18.5
beautifulsoup4==4.8.2
boto3==1.35.36
botocore==1.35.36
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
compressed-rtf==1.0.6
contourpy==1.3.1
cryptography==43.0.3
cssselect==1.2.0
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
ebcdic==1.1.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
extract-msg==0.28.7
faiss-cpu==1.9.0
fake-useragent==1.5.1
fastapi==0.115.5
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
fuzzywuzzy==0.18.0
gensim==4.3.3
graspologic==3.4.1
graspologic-native==1.2.1
greenlet==3.1.1
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
hyppo==0.4.0
idna==3.10
IMAPClient==2.1.0
importlib_metadata==8.5.0
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpickle==4.0.0
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langsmith==0.1.143
Levenshtein==0.26.1
lightrag==0.1.0b6
-e git+https://github.com/HKUDS/LightRAG.git@b00f2d602739f19c8c42bfd319ddc05c795d8355#egg=lightrag_hku
llvmlite==0.43.0
lxml==5.3.0
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nano-vectordb==0.0.4.3
neo4j==5.26.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
ollama==0.4.0
openai==1.54.4
openpyxl==3.1.5
oracledb==2.5.0
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parse==1.20.2
parso==0.8.4
pathlib==1.0.1
patsy==1.0.1
pdfminer.six==20191110
pexpect==4.9.0
pillow==11.0.0
POT==0.9.5
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyee==11.1.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.0
pypdf==5.1.0
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python-multipart==0.0.17
python-pptx==0.6.23
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
requests-toolbelt==1.0.0
rich==13.9.4
s3transfer==0.10.4
safetensors==0.4.5
scheduler==0.8.7
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
shellingham==1.5.4
six==1.12.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.0.0
textract==1.6.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
typeguard==4.4.1
typer==0.13.1
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
umap-learn==0.5.7
urllib3==1.26.20
uvicorn==0.32.0
uvloop==0.21.0
w3lib==2.2.1
watchfiles==0.24.0
wcwidth==0.2.13
websockets==10.4
wrapt==1.17.0
xlrd==1.2.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.17.2
zipp==3.21.0
